--- src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/events/WorkEvents.kt
@@ -1,15 +1,0 @@
-package slatekit.jobs.events
-
-import slatekit.functions.policy.Policy
-import slatekit.jobs.WorkRequest
-import slatekit.jobs.WorkState
-import slatekit.jobs.Worker
-
-interface WorkEvents : Events<Worker<*>> {
-
-    /**
-     * Applies the policy to these workers
-     * Policies add behaviour to workers such as retries, limits, error ratios, etc
-     */
-    fun apply(policy: Policy<WorkRequest, WorkState>)
-}
